Cit 18 harvrest detail management

+++ src/main/java/com/wora/citronix/tree/application/dto/TreeResponseDto.java
@@ -9,6 +9,6 @@ import java.time.LocalDate;
 public record TreeResponseDto(@NotNull Long id,
                               @NotNull LocalDate plantingDate,
                               @NotNull Level level,
-                              @NotNull Double age,
+                              @NotNull Integer age,
                               @NotNull FieldEmbeddableDto field) {
 }
